feat(attritionTableHistogramModal): Removed duplicate modal and hoisted existing modal to be a child of AttritionTableWrapper to try to resolve issue with outcome diagram not re-drawing
