+++ src/Analysis/GWASApp/Components/AttritionTableWrapper/AttritionTableWrapper.jsx
@@ -1,16 +1,38 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import AttritionTable from './AttritionTable/AttritionTable';
+import AttritionTableModal from './AttritionTableModal/AttritionTableModal';
 
 const AttritionTableWrapper = ({ covariates, selectedCohort, outcome }) => {
   const useSecondTable = outcome?.variable_type === 'custom_dichotomous';
+
+  const [modalInfo, setModalInfo] = useState({
+    title: '',
+    isModalOpen: false,
+    selectedCohort: null,
+    currentCovariateAndCovariatesFromPrecedingRows: null,
+    outcome: null,
+    rowObject: null,
+  });
+  // Keep modal info up-to-date with changes in the data needed for data viz
+  useEffect(() => {
+    setModalInfo({
+      ...modalInfo,
+      selectedCohort,
+      outcome,
+    });
+  }, [selectedCohort, covariates, outcome]);
+
   return (
     <div data-tour='attrition-table'>
+      <AttritionTableModal modalInfo={modalInfo} setModalInfo={setModalInfo} />
       <AttritionTable
         covariates={covariates}
         selectedCohort={selectedCohort}
         outcome={outcome}
         tableType={useSecondTable ? 'Case Cohort' : ''}
+        modalInfo={modalInfo}
+        setModalInfo={setModalInfo}
       />
       {useSecondTable && (
         <AttritionTable
@@ -18,6 +40,8 @@ const AttritionTableWrapper = ({ covariates, selectedCohort, outcome }) => {
           selectedCohort={selectedCohort}
           outcome={outcome}
           tableType={'Control Cohort'}
+          modalInfo={modalInfo}
+          setModalInfo={setModalInfo}
         />
       )}
     </div>
